added response entity as a common response for all apis

+++ src/main/java/org/girish/myfirstproject/Controller/JournalController.java
@@ -3,18 +3,16 @@ package org.girish.myfirstproject.Controller;
 import org.girish.myfirstproject.Model.JournalEntry;
 import org.girish.myfirstproject.Service.JournalService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 @RestController
 public class JournalController {
 
-    private Map<Long, JournalEntry> journalEntries = new HashMap<>();
     @Autowired
     JournalService journalService;
+
     @GetMapping("/")
     public String sayHello() {
         return "Hello World";
@@ -22,23 +20,44 @@ public class JournalController {
 
     @PostMapping("/add") // Changed from PUT to POST
     public String createJournal(@RequestBody JournalEntry journalEntry) {
-       return journalService.createEntry(journalEntry);
+        return journalService.createEntry(journalEntry);
     }
+
     @GetMapping("/allJournals")
-    public List<JournalEntry> getAllJournals() {
-        return journalService.getAllJournals();
+    public ResponseEntity<?> getAllJournals() {
+        if (journalService.getAllJournals() != null) {
+            return ResponseEntity.ok(journalService.getAllJournals());
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
+
     @GetMapping("/getJournalById/{id}")
-    public JournalEntry getJournalById(@PathVariable Long id) {
-        return journalService.getJournalById(id);
+    public ResponseEntity<?> getJournalById(@PathVariable Long id) {
+        if (journalService.getJournalById(id) != null) {
+            return ResponseEntity.ok(journalService.getJournalById(id));
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
+
     @DeleteMapping("/deleteJournalById/{id}")
-    public boolean deleteJournalById(@PathVariable Long id) {
-        journalService.deleteJournalById(id);
-       return true;
+    public ResponseEntity<?> deleteJournalById(@PathVariable Long id) {
+        if (journalService.getJournalById(id) != null) {
+            journalService.deleteJournalById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
+
     @PutMapping("/updateJournalById/{id}")
-    public JournalEntry updateJournalById(@PathVariable Long id, @RequestBody JournalEntry journalEntry) {
-        return journalService.updateJournalById(id,journalEntry);
+    public ResponseEntity<?> updateJournalById(@PathVariable Long id, @RequestBody JournalEntry journalEntry) {
+        if (journalService.getJournalById(id) != null) {
+            journalService.updateJournalById(id, journalEntry);
+            return ResponseEntity.ok(journalService.getJournalById(id));
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
